+++ NEWS.md
@@ -5,6 +5,7 @@
   [#354](https://github.com/lgatto/MSnbase/issues/354)) <2018-11-07 Wed>
 - Add as,Spectra,list as,Spectra,MSnExp and as,MSnExp,Spectra
   (issue [#370](https://github.com/lgatto/MSnbase/issues/370)) <2018-11-07 Wed>
+- Add filterMsLevel,Spectra <2018-11-07 Wed>
 
 ## Changes in version 2.9.0
 - New version for Bioc 3.9 devel